Move javaSecuritySecureRandom to SecureRandomCompanionPlatform
