+++ std/shared/src/main/scala/cats/effect/std/SecureRandom.scala
@@ -32,9 +32,7 @@ import cats.effect.kernel._
 import cats.effect.std.Random.{ScalaRandom, TranslatedRandom}
 import cats.syntax.all._
 
-import scala.util.{Random => SRandom, Try}
-
-import java.security.{SecureRandom => JavaSecureRandom}
+import scala.util.{Random => SRandom}
 
 /**
  * SecureRandom is the ability to get cryptographically strong random information. It is an
@@ -126,36 +124,6 @@ object SecureRandom extends SecureRandomCompanionPlatform {
     }
 
   /**
-   * Create a non-blocking Random instance. This is more efficient than [[javaSecurityRandom]],
-   * but requires a more thoughtful instance of random
-   *
-   * @param random
-   *   a non-blocking instance of java.util.Random. It is the caller's responsibility to assure
-   *   that the instance is non-blocking. If the instance blocks during seeding, it is the
-   *   caller's responsibility to seed it.
-   */
-  private def javaUtilRandomNonBlocking[F[_]: Sync](
-      random: JavaSecureRandom): F[SecureRandom[F]] =
-    Sync[F]
-      .delay(random)
-      .map(r => new ScalaRandom[F](Applicative[F].pure(r)) with SecureRandom[F] {})
-
-  /**
-   * Creates a blocking Random instance. All calls to nextBytes are shifted to the blocking
-   * pool. This is safer, but less effecient, than [[javaUtilRandomNonBlocking]].
-   *
-   * @param random
-   *   a potentially blocking instance of java.util.Random
-   */
-  private def javaUtilRandomBlocking[F[_]: Sync](random: JavaSecureRandom): F[SecureRandom[F]] =
-    Sync[F]
-      .blocking(random)
-      .map(r => new ScalaRandom[F](Applicative[F].pure(r)) with SecureRandom[F] {})
-
-  /**
-   * Creates a SecureRandom instance. On most platforms, it will be non-blocking. If a
-   * non-blocking instance can't be guaranteed, falls back to a blocking implementation.
-   *
    * On the JVM, delegates to [[java.security.SecureRandom]].
    *
    * In browsers, delegates to the
@@ -168,62 +136,8 @@ object SecureRandom extends SecureRandomCompanionPlatform {
    * on Linux, macOS, and BSD. Unsupported platforms such as Windows will encounter link-time
    * errors.
    */
-  def javaSecuritySecureRandom[F[_]: Sync]: F[SecureRandom[F]] = {
-    // This is a known, non-blocking, threadsafe algorithm
-    def happyRandom = Sync[F].delay(JavaSecureRandom.getInstance("NativePRNGNonBlocking"))
-
-    def fallback = Sync[F].delay(new JavaSecureRandom())
-
-    // Porting JavaSecureRandom.isThreadSafe
-    def isThreadsafe(rnd: JavaSecureRandom) =
-      rnd
-        .getProvider
-        .getProperty("SecureRandom." + rnd.getAlgorithm + " ThreadSafe", "false")
-        .toBoolean
-
-    // If we can't sniff out a more optimal solution, we can always
-    // fall back to a pool of blocking instances
-    def fallbackPool: F[SecureRandom[F]] =
-      for {
-        n <- Sync[F].delay(Runtime.getRuntime.availableProcessors())
-        sr <- javaSecuritySecureRandom(n)
-      } yield sr
-
-    javaMajorVersion
-      .flatMap {
-        case Some(major) if major > 8 =>
-          happyRandom.redeemWith(
-            _ =>
-              fallback.flatMap {
-                case rnd if isThreadsafe(rnd) =>
-                  // We avoided the mutex, but not the blocking.  Use a
-                  // shared instance from the blocking pool.
-                  javaUtilRandomBlocking(rnd)
-                case _ =>
-                  // We can't prove the instance is threadsafe, so we need
-                  // to pessimistically fall back to a pool.  This should
-                  // be exceedingly uncommon.
-                  fallbackPool
-              },
-            // We are thread safe and non-blocking.  This is the
-            // happy path, and happily, the common path.
-            rnd => javaUtilRandomNonBlocking(rnd)
-          )
-
-        case Some(_) | None =>
-          // We can't guarantee we're not stuck in a mutex.
-          fallbackPool
-      }
+  def javaSecuritySecureRandom[F[_]: Sync]: F[SecureRandom[F]] =
+    Sync[F]
+      .delay(new JavaSecureRandom())
       .map(r => new ScalaRandom[F](Applicative[F].pure(r)) with SecureRandom[F] {})
-  }
-
-  private def javaMajorVersion[F[_]: Sync]: F[Option[Int]] =
-    Sync[F].delay(sys.props.get("java.version")).map(_.flatMap(parseJavaMajorVersion))
-
-  private def parseJavaMajorVersion(javaVersion: String): Option[Int] =
-    if (javaVersion.startsWith("1."))
-      Try(javaVersion.split("\\.")(1).toInt).toOption
-    else
-      Try(javaVersion.split("\\.")(0).toInt).toOption
-
 }
